make softfloat cast portable

+++ nbre/runtime/dip/dip_reward.cpp
@@ -59,7 +59,7 @@ std::unique_ptr<std::vector<dip_info_t>> dip_reward::get_dip_reward(
   // bonus_total = adb_ptr->get_normalized_value(bonus_total);
 
   floatxx_t::value_type zero =
-      softfloat_cast<uint32_t, typename floatxx_t::value_type>(0);
+      softfloat_cast<uint32_t, typename floatxx_t::value_type>(uint32_t(0));
   floatxx_t sum_votes(zero);
   for (auto &v : *it_dapp_votes) {
     sum_votes += v.second * v.second;
@@ -205,7 +205,7 @@ dip_reward::account_to_contract_votes(
     }
 
     floatxx_t::value_type zero =
-        softfloat_cast<uint32_t, typename floatxx_t::value_type>(0);
+        softfloat_cast<uint32_t, typename floatxx_t::value_type>(uint32_t(0));
     floatxx_t sum_votes(zero);
     for (auto &e : it_acc->second) {
       sum_votes += e.second;
@@ -249,9 +249,9 @@ floatxx_t dip_reward::participate_lambda(
     const std::vector<neb::rt::nr::nr_info_t> &nr_infos) {
 
   floatxx_t::value_type zero =
-      softfloat_cast<uint32_t, typename floatxx_t::value_type>(0);
+      softfloat_cast<uint32_t, typename floatxx_t::value_type>(uint32_t(0));
   floatxx_t::value_type one =
-      softfloat_cast<uint32_t, typename floatxx_t::value_type>(1);
+      softfloat_cast<uint32_t, typename floatxx_t::value_type>(uint32_t(1));
 
   std::unordered_set<std::string> addr_set;
   for (auto &tx : txs) {
